1.19 update and pages

+++ shared/Types/BeatSaver/UserStats.hpp
@@ -9,7 +9,7 @@ DECLARE_JSON_CLASS(BeatSaver, UserStats,
     GETTER_VALUE(int, RankedMaps);
     GETTER_VALUE(float, AvgBpm);
     GETTER_VALUE(float, AvgScore);
-    GETTER_VALUE(int, AvgDuration);
+    GETTER_VALUE(float, AvgDuration);
     GETTER_VALUE(std::string, FirstUpload);
     GETTER_VALUE(std::string, LastUpload);
     GETTER_VALUE(BeatSaver::UserDiffStats, DiffStats);
